backend/requirements.txt
fastapi
uvicorn
yt-dlp
python-dotenv

   